NoUploadStatsToController to not send to controller

+++ pkg/pillar/hypervisor/kubevirt.go
@@ -72,6 +72,7 @@ type vmiMetaData struct {
 	cputotal         uint64                               // total CPU in NS so far
 	maxmem           uint32                               // total Max memory usage in bytes so far
 	startUnknownTime time.Time                            // time when the domain returned as unknown status
+	scheduledLocal   bool                                 // if the domain is scheduled to local node
 }
 
 type kubevirtContext struct {
@@ -95,6 +96,7 @@ var stateMap = map[string]types.SwState{
 	"Scheduling": types.SCHEDULING,
 	"Failed":     types.FAILED,
 	"Halting":    types.HALTING,
+	"Succeeded":  types.SCHEDULING,
 	"Unknown":    types.UNKNOWN,
 }
 
@@ -584,16 +586,21 @@ func (ctx kubevirtContext) Stop(domainName string, force bool) error {
 	if !ok {
 		return logError("domain %s failed to get vmlist", domainName)
 	}
-	if vmis.mtype == IsMetaReplicaPod {
-		err = StopReplicaPodContainer(kubeconfig, vmis.name)
-	} else if vmis.mtype == IsMetaReplicaVMI {
-		err = StopReplicaVMI(kubeconfig, vmis.name)
-	} else {
-		return logError("Stop domain %s wrong type", domainName)
-	}
+	// call .Info() to decide if the domain is running on local node
+	_, _, err = ctx.Info(domainName)
+	logrus.Infof("Stop domain %s, scheduledLocal %v", domainName, vmis.scheduledLocal)
+	if vmis.scheduledLocal {
+		if vmis.mtype == IsMetaReplicaPod {
+			err = StopReplicaPodContainer(kubeconfig, vmis.name)
+		} else if vmis.mtype == IsMetaReplicaVMI {
+			err = StopReplicaVMI(kubeconfig, vmis.name)
+		} else {
+			return logError("Stop domain %s wrong type", domainName)
+		}
 
-	if err != nil {
-		return err
+		if err != nil {
+			return err
+		}
 	}
 
 	delete(ctx.vmiList, domainName)
@@ -615,16 +622,22 @@ func (ctx kubevirtContext) Delete(domainName string) (result error) {
 	if !ok {
 		return logError("delete domain %s failed to get vmlist", domainName)
 	}
-	if vmis.mtype == IsMetaReplicaPod {
-		err = StopReplicaPodContainer(kubeconfig, vmis.name)
-	} else if vmis.mtype == IsMetaReplicaVMI {
-		err = StopReplicaVMI(kubeconfig, vmis.name)
-	} else {
-		return logError("delete domain %s wrong type", domainName)
-	}
 
-	if err != nil {
-		return err
+	// call .Info() to decide if the domain is running on local node
+	_, _, err = ctx.Info(domainName)
+	logrus.Infof("Delete domain %s, scheduledLocal %v", domainName, vmis.scheduledLocal)
+	if vmis.scheduledLocal {
+		if vmis.mtype == IsMetaReplicaPod {
+			err = StopReplicaPodContainer(kubeconfig, vmis.name)
+		} else if vmis.mtype == IsMetaReplicaVMI {
+			err = StopReplicaVMI(kubeconfig, vmis.name)
+		} else {
+			return logError("delete domain %s wrong type", domainName)
+		}
+
+		if err != nil {
+			return err
+		}
 	}
 
 	// Delete the state dir
@@ -681,13 +694,20 @@ func (ctx kubevirtContext) Info(domainName string) (int, types.SwState, error) {
 		return 0, types.HALTED, logError("info domain %s failed to get vmlist", domainName)
 	}
 	if vmis.mtype == IsMetaReplicaPod {
-		res, err = InfoReplicaSetContainer(ctx, vmis)
+		res, err = InfoReplicaSetContainer(ctx, vmis, nodeName)
 	} else {
 		res, err = getVMIStatus(vmis, nodeName)
 	}
 
 	if effectiveDomainState, matched := stateMap[res]; !matched {
-		return 0, types.BROKEN, logError("domain %s reported to be in unexpected state %s", domainName, res)
+		// Received undefined state in our map, return UNKNOWN instead
+		retStatus, err := checkAndReturnStatus(vmis, true)
+		logrus.Infof("domain %s reported to be in unexpected state %s", domainName, res)
+		effectiveDomainState = types.HALTING
+		if retStatus == "Unknown" {
+			effectiveDomainState = types.UNKNOWN
+		}
+		return ctx.vmiList[domainName].domainID, effectiveDomainState, err
 	} else {
 		if _, ok := ctx.vmiList[domainName]; !ok { // domain is deleted
 			return 0, types.HALTED, logError("domain %s is deleted", domainName)
@@ -711,18 +731,24 @@ func (ctx kubevirtContext) Cleanup(domainName string) error {
 	if !ok {
 		return logError("cleanup domain %s failed to get vmlist", domainName)
 	}
-	if vmis.mtype == IsMetaReplicaPod {
-		_, err = InfoReplicaSetContainer(ctx, vmis)
-		if err == nil {
-			err = ctx.Delete(domainName)
+
+	// call .Info() to decide if the domain is running on local node
+	_, _, err = ctx.Info(domainName)
+	logrus.Infof("Cleanup domain %s, scheduledLocal %v", domainName, vmis.scheduledLocal)
+	if vmis.scheduledLocal {
+		if vmis.mtype == IsMetaReplicaPod {
+			_, err = InfoReplicaSetContainer(ctx, vmis, nodeName)
+			if err == nil {
+				err = ctx.Delete(domainName)
+			}
+		} else if vmis.mtype == IsMetaReplicaVMI {
+			err = waitForVMI(vmis, nodeName, false)
+		} else {
+			err = logError("cleanup domain %s wrong type", domainName)
 		}
-	} else if vmis.mtype == IsMetaReplicaVMI {
-		err = waitForVMI(vmis, nodeName, false)
-	} else {
-		err = logError("cleanup domain %s wrong type", domainName)
-	}
-	if err != nil {
-		return fmt.Errorf("waitforvmi failed  %s: %v", domainName, err)
+		if err != nil {
+			return fmt.Errorf("waitforvmi failed  %s: %v", domainName, err)
+		}
 	}
 
 	return nil
@@ -779,11 +805,13 @@ func getVMIStatus(vmis *vmiMetaData, nodeName string) (string, error) {
 		if vmi.Status.NodeName == nodeName {
 			if vmi.GenerateName == repVmiName {
 				targetVMI = &vmi
+				vmis.scheduledLocal = true
 				break
 			}
 		} else {
 			if vmi.GenerateName == repVmiName {
 				foundNonlocal = true
+				vmis.scheduledLocal = false
 				nonLocalStatus = fmt.Sprintf("%v", vmi.Status.Phase)
 			}
 		}
@@ -1296,12 +1324,18 @@ func checkForReplicaPod(ctx kubevirtContext, vmis *vmiMetaData) error {
 	var i int
 	var status string
 	var err error
+
+	nodeName, ok := ctx.nodeNameMap["nodename"]
+	if !ok {
+		return logError("Cleanup: Failed to get nodeName")
+	}
+
 	for {
 		i++
 		logrus.Infof("checkForReplicaPod: check(%d) wait 15 sec, %v", i, repName)
 		time.Sleep(15 * time.Second)
 
-		status, err = InfoReplicaSetContainer(ctx, vmis)
+		status, err = InfoReplicaSetContainer(ctx, vmis, nodeName)
 		if err != nil {
 			logrus.Infof("checkForReplicaPod: repName %s, %v", repName, err)
 		} else {
@@ -1320,7 +1354,7 @@ func checkForReplicaPod(ctx kubevirtContext, vmis *vmiMetaData) error {
 	return fmt.Errorf("checkForReplicaPod: timed out, statuus %s, err %v", status, err)
 }
 
-func InfoReplicaSetContainer(ctx kubevirtContext, vmis *vmiMetaData) (string, error) {
+func InfoReplicaSetContainer(ctx kubevirtContext, vmis *vmiMetaData, nodeName string) (string, error) {
 
 	repName := vmis.repPod.ObjectMeta.Name
 	err := getConfig(&ctx)
@@ -1343,6 +1377,11 @@ func InfoReplicaSetContainer(ctx kubevirtContext, vmis *vmiMetaData) (string, er
 	}
 
 	for _, pod := range pods.Items {
+		if pod.Spec.NodeName != nodeName {
+			vmis.scheduledLocal = false
+		} else {
+			vmis.scheduledLocal = true
+		}
 
 		var res string
 		// https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle/
